+++ app/Models/Config.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright (c) 2018-2021 Adshares sp. z o.o.
+ * Copyright (c) 2018-2022 Adshares sp. z o.o.
  *
  * This file is part of AdServer
  *
@@ -28,11 +28,9 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Cache;
 
-use function array_merge;
-use function in_array;
-use function sprintf;
-
 /**
+ * @property string key
+ * @property string value
  * @mixin Builder
  */
 class Config extends Model
@@ -77,6 +75,16 @@ class Config extends Model
     public const INVOICE_COMPANY_COUNTRY = 'invoice-company-country';
     public const INVOICE_COMPANY_VAT_ID = 'invoice-company-vat-id';
     public const INVOICE_COMPANY_BANK_ACCOUNTS = 'invoice-company-bank-accounts';
+    public const SITE_ACCEPT_BANNERS_MANUALLY = 'site-accept-banners-manually';
+    public const SITE_CLASSIFIER_LOCAL_BANNERS = 'site-classifier-local-banners';
+    public const ALLOWED_CLASSIFIER_LOCAL_BANNERS_OPTIONS = [
+        Config::CLASSIFIER_LOCAL_BANNERS_ALL_BY_DEFAULT,
+        Config::CLASSIFIER_LOCAL_BANNERS_LOCAL_BY_DEFAULT,
+        Config::CLASSIFIER_LOCAL_BANNERS_LOCAL_ONLY,
+    ];
+    public const CLASSIFIER_LOCAL_BANNERS_ALL_BY_DEFAULT = 'all-by-default';
+    public const CLASSIFIER_LOCAL_BANNERS_LOCAL_BY_DEFAULT = 'local-by-default';
+    public const CLASSIFIER_LOCAL_BANNERS_LOCAL_ONLY = 'local-only';
 
     private const ADMIN_SETTINGS_DEFAULTS = [
         self::OPERATOR_TX_FEE => '',